Conditonally import tolerant comparator per platform in rfw's material_widget_test.dart
