+++ packages/rfw/test/tolerant_comparator_web.dart
@@ -1,0 +1,6 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+void setUpTolerantComparator(
+    {required String testPath, required double precisionTolerance}) {}
