+++ packages/rfw/test/utils.dart
@@ -5,7 +5,6 @@
 import 'dart:io' show Platform;
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter_test/flutter_test.dart';
 
 // Detects if we're running the tests on the main channel.
 //
@@ -22,55 +21,3 @@ bool get isMainChannel {
 
 // See Contributing section of README.md file.
 final bool runGoldens = !kIsWeb && Platform.isLinux && isMainChannel;
-
-/// Sets [_TolerantGoldenFileComparator] as the default golden file comparator
-/// in tests.
-void setUpTolerantComparator(
-    {required String testPath, required double precisionTolerance}) {
-  if (!kIsWeb) {
-    final GoldenFileComparator oldComparator = goldenFileComparator;
-    final _TolerantGoldenFileComparator newComparator =
-        _TolerantGoldenFileComparator(Uri.parse(testPath),
-            precisionTolerance: precisionTolerance);
-
-    goldenFileComparator = newComparator;
-
-    addTearDown(() => goldenFileComparator = oldComparator);
-  }
-}
-
-class _TolerantGoldenFileComparator extends LocalFileComparator {
-  _TolerantGoldenFileComparator(
-    super.testFile, {
-    required double precisionTolerance,
-  })  : assert(
-          0 <= precisionTolerance && precisionTolerance <= 1,
-          'precisionTolerance must be between 0 and 1',
-        ),
-        _precisionTolerance = precisionTolerance;
-
-  /// How much the golden image can differ from the test image.
-  ///
-  /// It is expected to be between 0 and 1. Where 0 is no difference (the same image)
-  /// and 1 is the maximum difference (completely different images).
-  final double _precisionTolerance;
-
-  @override
-  Future<bool> compare(Uint8List imageBytes, Uri golden) async {
-    final ComparisonResult result = await GoldenFileComparator.compareLists(
-      imageBytes,
-      await getGoldenBytes(golden),
-    );
-
-    final bool passed =
-        result.passed || result.diffPercent <= _precisionTolerance;
-    if (passed) {
-      result.dispose();
-      return true;
-    }
-
-    final String error = await generateFailureOutput(result, golden, basedir);
-    result.dispose();
-    throw FlutterError(error);
-  }
-}
